requirements.txt
kafka-python
flake8